add helper class, fixed some lines
